HOTFIX added Scalar to Answer for UI Rating Implementation
Issue is that for Rating Questions the Answer Item needs the scalar since it is not present in the UI Data Model.

+++ src/cloud/backend/apollo.cloud/graph.invite-apollo.app.common/Models/UserProfile/AssessmentCategoryResult.cs
@@ -25,7 +25,7 @@ namespace Invite.Apollo.App.Graph.Common.Models.UserProfile
 
         //The result a user scored in a Category
         [DataMember(Order = 7)]
-        public long Result { get; set; }
+        public decimal Result { get; set; }
 
         [ForeignKey(nameof(CourseItem))]
         [DataMember(Order = 8)]
